+++ src/beekeeper/__init__.py
@@ -9,6 +9,7 @@ from beekeeper.adapters.inputs.mixed_input_adapter import MixedInputAdapter
 from beekeeper.adapters.outputs.output_adapter import OutputAdapter
 from beekeeper.algorithm.algorithm_state import State
 from beekeeper.algorithm.base_algorithm import BaseAlgorithm
+from beekeeper.algorithm.errors import IncompleteSolutionError
 from beekeeper.allocations.allocation_request import AllocationRequest
 from beekeeper.allocations.allocation_type import AllocationType
 from beekeeper.allocations.planned_allocation import PlannedAllocation
@@ -32,6 +33,7 @@ __all__ = [
     "HardPreliminaryRule",
     "HardStatefulRule",
     "Inavailability",
+    "IncompleteSolutionError",
     "InputAdapter",
     "JsonAllocationInputAdapter",
     "JsonEntityInputAdapter",